+++ Game/src/audiolib/pitch.h
@@ -31,6 +31,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #ifndef __PITCH_H
 #define __PITCH_H
 
+#include "../duke3d.h"
+
 enum PITCH_ERRORS
    {
    PITCH_Warning = -2,
@@ -39,7 +41,7 @@ enum PITCH_ERRORS
    };
 
 //void          PITCH_Init( void );
-unsigned long PITCH_GetScale( int pitchoffset );
+uint32_t PITCH_GetScale( int pitchoffset );
 void          PITCH_UnlockMemory( void );
 int           PITCH_LockMemory( void );
 #endif
